--- a/cli/Alias.hx
+++ b/cli/Alias.hx
@@ -1,0 +1,9 @@
+package;
+
+
+class Alias {
+	public static function main () {
+		var args = [ "run", "haxeui-core" ].concat (Sys.args ());
+		Sys.exit (Sys.command ("haxelib", args));
+	}
+}