--- a/cli/Package.bat
+++ b/cli/Package.bat
@@ -1,0 +1,5 @@
+@echo off
+cd bin
+echo Creating standalone command line executable...
+nekotools boot ../run.n
+pause