--- a/cli/Run.bat
+++ b/cli/Run.bat
@@ -1,0 +1,3 @@
+@echo off
+neko ../run.n
+pause