--- a/cli/alias.hxml
+++ b/cli/alias.hxml
@@ -1,0 +1,3 @@
+-main Alias
+-neko haxeui.n
+-cmd nekotools boot haxeui.n