--- a/cli/haxeui-tools.hxproj
+++ b/cli/haxeui-tools.hxproj
@@ -24,7 +24,7 @@
     <option noInlineOnDebug="False" />
     <option mainClass="Main" />
     <option enabledebug="False" />
-    <option additional="-resource projects/shared/project.json@projects/shared/project.json&#xA;-resource projects/html5/project.json@projects/html5/project.json&#xA;-resource projects/html5-flash-develop/project.json@projects/html5-flash-develop/project.json&#xA;-resource projects/hxwidgets/project.json@projects/hxwidgets/project.json&#xA;-resource projects/openfl/project.json@projects/openfl/project.json" />
+    <option additional="-resource projects/shared/project.json@projects/shared/project.json&#xA;-resource projects/html5/project.json@projects/html5/project.json&#xA;-resource projects/html5-flash-develop/project.json@projects/html5-flash-develop/project.json&#xA;-resource projects/hxwidgets/project.json@projects/hxwidgets/project.json&#xA;-resource projects/openfl/project.json@projects/openfl/project.json&#xA;-resource projects/nme/project.json@projects/nme/project.json" />
   </build>
   <!-- haxelib libraries -->
   <haxelib>