--- a/cli/src/Main.hx
+++ b/cli/src/Main.hx
@@ -36,34 +36,40 @@ class Main {
         }
         
         if (command == "setup") {
-            // almost certainly a much better way to be doing this!
-            var haxeLibPath = Sys.getCwd(); //HaxeLibHelper.getLibPath("haxeui-core");
-            
-            log('Setting up haxeui tools from "${haxeLibPath}"');
-            File.copy('${haxeLibPath}/cli/Alias.hx', 'Alias.hx');
-            File.copy('${haxeLibPath}/cli/alias.hxml', 'alias.hxml');
-            
-            log("Building haxeui alias");
-            Sys.command("haxe", ['alias.hxml']);
-            
             var haxePath = Sys.getEnv("HAXEPATH");
-            log('Copying alias to "${haxePath}"');
-            if (FileSystem.exists('haxeui.exe')) {
-                File.copy('haxeui.exe', '${haxePath}/haxeui.exe');
-            }
-            if (FileSystem.exists('haxeui')) {
-                File.copy('haxeui', '${haxePath}/haxeui');
-            }
-            
-            log('Cleaning up');
-            FileSystem.deleteFile("Alias.hx");
-            FileSystem.deleteFile("alias.hxml");
-            FileSystem.deleteFile("haxeui.n");
-            if (FileSystem.exists('haxeui.exe')) {
-                FileSystem.deleteFile("haxeui.exe");
-            }
-            if (FileSystem.exists('haxeui')) {
-                FileSystem.deleteFile("haxeui");
+            if (haxePath != null) {
+                // almost certainly a much better way to be doing this!
+                var haxeLibPath = Sys.getCwd(); //HaxeLibHelper.getLibPath("haxeui-core");
+                
+                log('Setting up haxeui tools from "${haxeLibPath}"');
+                File.copy('${haxeLibPath}/cli/Alias.hx', 'Alias.hx');
+                File.copy('${haxeLibPath}/cli/alias.hxml', 'alias.hxml');
+                
+                log("Building haxeui alias");
+                Sys.command("haxe", ['alias.hxml']);
+                
+                log('Copying alias to "${haxePath}"');
+                if (FileSystem.exists('haxeui.exe')) {
+                    File.copy('haxeui.exe', '${haxePath}/haxeui.exe');
+                }
+                if (FileSystem.exists('haxeui')) {
+                    File.copy('haxeui', '${haxePath}/haxeui');
+                }
+                
+                log('Cleaning up');
+                FileSystem.deleteFile("Alias.hx");
+                FileSystem.deleteFile("alias.hxml");
+                FileSystem.deleteFile("haxeui.n");
+                if (FileSystem.exists('haxeui.exe')) {
+                    FileSystem.deleteFile("haxeui.exe");
+                }
+                if (FileSystem.exists('haxeui')) {
+                    FileSystem.deleteFile("haxeui");
+                }
+            } else { // more hacky!
+                log("Setting up alias (may need sudo)");
+                var content = "#!/bin/sh\nhaxelib run haxeui-core \"$@\"";
+                File.saveContent("/usr/local/bin/haxeui", content);
             }
             return;
         }