--- a/cli/src/Main.hx
+++ b/cli/src/Main.hx
@@ -2,6 +2,7 @@ package;
 
 import commands.CommandFactory;
 import projects.Project;
+import projects.ProjectFactory;
 import sys.FileSystem;
 import sys.io.File;