--- a/cli/src/Params.hx
+++ b/cli/src/Params.hx
@@ -1,0 +1,9 @@
+package;
+
+typedef Params = {
+    cwd: String,
+    target: String,
+    command: String,
+    backend: String,
+    ?additional: Array<String>
+}