--- a/cli/src/builds/Build.hx
+++ b/cli/src/builds/Build.hx
@@ -1,0 +1,11 @@
+package builds;
+
+class Build {
+    public function new() {
+        
+    }
+    
+    public function execute(params:Params) {
+        
+    }
+}