… cli/haxeui.n
… cli/projects/html5-flash-develop/project.json
… cli/projects/html5-shared/project.json
… cli/projects/html5/project.json
… cli/projects/hxwidgets-flash-develop/project.json
… cli/projects/hxwidgets-shared/project.json
… cli/projects/hxwidgets/project.json
… cli/projects/kha-shared/project.json
… cli/projects/kha/project.json
… cli/projects/nme-flash-develop/project.json
… cli/projects/nme-shared/project.json
… cli/projects/nme/project.json
… cli/projects/openfl-flash-develop/project.json
… cli/projects/openfl-shared/project.json
… cli/projects/openfl/project.json
… cli/projects/pixijs-flash-develop/project.json
… cli/projects/pixijs-shared/project.json
… cli/projects/pixijs/project.json
… cli/projects/shared/project.json
… cli/src/builds/BuildFactory.hx
… cli/src/builds/ProcessBuild.hx
… cli/src/builds/html5/Html5Build.hx
… cli/src/builds/hxwidgets/HxWidgetsBuild.hx
… cli/src/builds/kha/KhaBuild.hx
… cli/src/builds/nme/NmeBuild.hx
… cli/src/builds/openfl/OpenFLBuild.hx
… cli/src/builds/pixijs/PixiJsBuilder.hx
… cli/src/commands/BuildCommand.hx
… cli/src/commands/Command.hx
… cli/src/commands/CommandFactory.hx