New validation system based in Feathers. Work in progress. Feedback welcome.

### Tests

A test with HTML 5 native. It adds 1000 buttons with label + icon in a Box with DefaultLayout:

Old:
![old](https://cloud.githubusercontent.com/assets/1063719/23145651/3464fa02-f7d0-11e6-8230-9a54759bde5b.png)


New:
![new](https://cloud.githubusercontent.com/assets/1063719/23145655/36cbe490-f7d0-11e6-80e2-d272c34baee1.png)

### Results

* Scripting: 8762 ms vs 2259 ms . 300% faster
* Rendering: 3754 ms vs 1852 ms. 250% faster

### Backends implementation

* [ ] HTML5 https://github.com/haxeui/haxeui-html5/pull/14
* [ ] OpenFL https://github.com/haxeui/haxeui-openfl/pull/13